Added more support for cmdline args
I added support for the basic --help and --version cmdline args as well
as the --text and --gui options for my game that allow you to boot into
text or gui mode.

The game is now booted by a seperate load game method which both
calculates which game to load and then loads it.

I am hoping in the future to have the gui mode act as a sort of wrapper
for the text mode, translating the actions made in the gui to the text
mode equivalents but that could be a pain/interesting challenge.
I'll have to ask dad about it.

I also removed an ugly new line at the start of the game.