Copied in files from pg skeleton
I copied in the files from my pygame project skeleton so that I could
start to add the gui part. I made some name changes to files to stop them
from clashing (game to textGame & guiGame) and moved the pg main into the
existing main.

For how to join the two projects together Dad has suggested creating a library
with all the actions required for the game that the two front ends interact with.
This seems like a much better and more achievable idea than layering the gui
ontop of the text version. However, I haven't started with this yet.