Made minor changes
I made minor changes to some files for aesthetic purposes and also
started to plan the refactor into using multiple front ends. I found
that the library idea seems quite complex and that I may come back to it
later but for now I'm focused on getting the gui working.

I have also found some wierd things with how the deck file and class are
used. In the previous commit, the game had a self.deck variable that
both cpu and player would reference and it was also used in parts of the
code in textGame. However, this seemed strange to me so I instead now
have the player and cpu with their own deck variables. The name of deck
also seems strange as currently it is more of a library of cards owned
than a deck of cards on a table. This also causes name clashes with
the player.deck and cpu.deck variables.